+++ .credo.exs
@@ -41,7 +41,7 @@
       # If you create your own checks, you must specify the source files for
       # them here, so they can be loaded by Credo before running the analysis.
       #
-      requires: [],
+      requires: ["./lib/skuld/credo/comp_pure_redundant.ex"],
       #
       # If you want to enforce a style guide and need a more traditional linting
       # experience, you can change `strict` to `true` below:
@@ -115,6 +115,7 @@
         #
         ## Refactoring Opportunities
         #
+        {Skuld.Credo.CompPureRedundant, []},
         {Credo.Check.Refactor.CondStatements, []},
         {Credo.Check.Refactor.CyclomaticComplexity, []},
         {Credo.Check.Refactor.FunctionArity, [max_arity: 10]},